task_api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pytest
httpx
python-dotenv
psycopg2